fix: implement King's suggestions
- move creating the event to .../attributes
- move .../event/create/general-info to .../event/create
