+++ src/app/dashboard/event/create/attributes/page.tsx
@@ -2,14 +2,13 @@
 
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useAtom } from "jotai";
-import { ArrowRight, Loader2, PlusIcon, TextIcon } from "lucide-react";
-import Link from "next/link";
+import { Loader2, PlusIcon, TextIcon } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
-import { Button, buttonVariants } from "@/components/ui/button";
+import { Button } from "@/components/ui/button";
 import {
   Form,
   FormControl,
@@ -25,7 +24,9 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
+import { toast } from "@/hooks/use-toast";
 
+import { saveEvent } from "../actions";
 import { FormContainer } from "../form-container";
 import { AttributeTypes, eventAtom } from "../state";
 
@@ -51,9 +52,25 @@ export default function AttributesForm() {
     setAttributes([...attributes, data]);
     form.reset();
   }
-  function saveAttributes() {
-    setEvent({ ...event, attributes });
-    router.push("/dashboard/event/create");
+  async function createEvent() {
+    try {
+      const result = await saveEvent({ ...event, attributes });
+      if ("errors" in result) {
+        toast({
+          variant: "destructive",
+          title: "O nie! Coś poszło nie tak.",
+          description: "Spróbuj utworzyć wydarzenie ponownie.",
+        });
+      } else {
+        router.push(`/dashboard/event/share/${result.id}`);
+      }
+    } catch {
+      toast({
+        variant: "destructive",
+        title: "Brak połączenia z serwerem.",
+        description: "Sprawdź swoje połączenie z internetem.",
+      });
+    }
   }
   return (
     <FormContainer
@@ -62,104 +79,100 @@ export default function AttributesForm() {
       description="Dodaj atrybuty"
       icon={<TextIcon />}
     >
-      <div className="space-y-4">
+      <div className="flex w-full flex-col items-center">
         <div className="space-y-4">
-          <div className="space-y-2">
-            {attributes.length > 0 && (
-              <p className="text-sm font-medium">Atrybuty</p>
-            )}
-            {attributes.map((attribute, key) => (
-              <div key={key} className="flex flex-row items-center gap-2">
-                <p className="flex h-12 w-full rounded-xl border border-input bg-transparent px-4 py-3 text-lg shadow-sm transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium file:text-foreground placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring md:text-sm">
-                  {attribute.name}
-                </p>
-                <p className="flex h-12 w-full rounded-xl border border-input bg-transparent px-4 py-3 text-lg shadow-sm transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium file:text-foreground placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring md:text-sm">
-                  {attribute.type}
-                </p>
-              </div>
-            ))}
-          </div>
-          <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-2">
-              <FormLabel>Dodaj atrybut</FormLabel>
-              <div className="flex flex-row items-center gap-2">
-                <FormField
-                  name="name"
-                  control={form.control}
-                  render={({ field }) => (
-                    <Input
-                      {...field}
-                      placeholder="Nazwa"
-                      disabled={form.formState.isSubmitting}
-                    />
-                  )}
-                />
-                <FormField
-                  name="type"
-                  control={form.control}
-                  render={({ field }) => (
-                    <FormItem>
-                      <Select
-                        onValueChange={field.onChange}
-                        defaultValue={field.value}
-                      >
-                        <FormControl>
-                          <SelectTrigger
-                            disabled={form.formState.isSubmitting}
-                            className="h-full capitalize"
-                          >
-                            <SelectValue placeholder="Wybierz typ atrybutu" />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          {AttributeTypes.map((value) => (
-                            <SelectItem
-                              className="capitalize"
-                              key={value}
-                              value={value}
+          <div className="space-y-4">
+            <div className="space-y-2">
+              {attributes.length > 0 && (
+                <p className="text-sm font-medium">Atrybuty</p>
+              )}
+              {attributes.map((attribute, key) => (
+                <div key={key} className="flex flex-row items-center gap-2">
+                  <p className="flex h-12 w-full rounded-xl border border-input bg-transparent px-4 py-3 text-lg shadow-sm transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium file:text-foreground placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring md:text-sm">
+                    {attribute.name}
+                  </p>
+                  <p className="flex h-12 w-full rounded-xl border border-input bg-transparent px-4 py-3 text-lg shadow-sm transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium file:text-foreground placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring md:text-sm">
+                    {attribute.type}
+                  </p>
+                </div>
+              ))}
+            </div>
+            <Form {...form}>
+              <form
+                onSubmit={form.handleSubmit(onSubmit)}
+                className="space-y-2"
+              >
+                <FormLabel>Dodaj atrybut</FormLabel>
+                <div className="flex flex-row items-center gap-2">
+                  <FormField
+                    name="name"
+                    control={form.control}
+                    render={({ field }) => (
+                      <Input
+                        {...field}
+                        placeholder="Nazwa"
+                        disabled={form.formState.isSubmitting}
+                      />
+                    )}
+                  />
+                  <FormField
+                    name="type"
+                    control={form.control}
+                    render={({ field }) => (
+                      <FormItem>
+                        <Select
+                          onValueChange={field.onChange}
+                          defaultValue={field.value}
+                        >
+                          <FormControl>
+                            <SelectTrigger
+                              disabled={form.formState.isSubmitting}
+                              className="h-full capitalize"
                             >
-                              {value}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    </FormItem>
-                  )}
-                />
-                <Button
-                  disabled={form.formState.isSubmitting}
-                  type="submit"
-                  variant="outline"
-                >
-                  <PlusIcon />
-                </Button>
-              </div>
-            </form>
-          </Form>
-        </div>
-        <div className="flex flex-row items-center justify-between gap-4">
-          <Link
-            className={buttonVariants({ variant: "ghost" })}
-            href="/dashboard/event/create"
-          >
-            Może później
-          </Link>
-          <Button
-            className="w-min"
-            variant="ghost"
-            onClick={saveAttributes}
-            disabled={form.formState.isSubmitting}
-          >
-            {form.formState.isSubmitting ? (
-              <>
-                Zapisywanie danych... <Loader2 className="animate-spin" />
-              </>
-            ) : (
-              <>
-                Dalej <ArrowRight />
-              </>
-            )}
-          </Button>
+                              <SelectValue placeholder="Wybierz typ atrybutu" />
+                            </SelectTrigger>
+                          </FormControl>
+                          <SelectContent>
+                            {AttributeTypes.map((value) => (
+                              <SelectItem
+                                className="capitalize"
+                                key={value}
+                                value={value}
+                              >
+                                {value}
+                              </SelectItem>
+                            ))}
+                          </SelectContent>
+                        </Select>
+                      </FormItem>
+                    )}
+                  />
+                  <Button
+                    disabled={form.formState.isSubmitting}
+                    type="submit"
+                    variant="outline"
+                  >
+                    <PlusIcon />
+                  </Button>
+                </div>
+              </form>
+            </Form>
+          </div>
+          <div className="flex w-full flex-row justify-end">
+            <Button
+              className="w-min"
+              onClick={createEvent}
+              disabled={form.formState.isSubmitting}
+            >
+              {form.formState.isSubmitting ? (
+                <>
+                  Zapisywanie danych... <Loader2 className="animate-spin" />
+                </>
+              ) : (
+                <>Utwórz wydarzenie</>
+              )}
+            </Button>
+          </div>
         </div>
       </div>
     </FormContainer>
