+++ src/app/dashboard/page.tsx
@@ -8,7 +8,7 @@ export default function DashboardHomepage() {
     <div className="flex flex-col gap-4">
       <h1 className="text-3xl font-bold">Panel organizatora</h1>
       <Link
-        href="/dashboard/event/create/general-info"
+        href="/dashboard/event/create"
         className={buttonVariants({ variant: "default" })}
       >
         Click me <Smile />
